use new INamed interface, avoid cyclic deps
